requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.13.1
psycopg2==2.9.3
Pygments==2.11.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.0
